requirements.txt
aiogram==3.12.0
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
SQLAlchemy==2.0.32
alembic==1.13.2
apscheduler==3.10.4
uvloop==0.19.0; sys_platform == 'linux'
